Starts refactor of canonical sort ordering work
